% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipreader.R
\name{read.pileup}
\alias{read.pileup}
\title{Read chip-seq files
This function does something}
\usage{
read.pileup(pileup_path)
}
\description{
Read chip-seq files
This function does something
}
\examples{

}
